feat(v8): extract V8 range intersection into oxc_coverage_v8 crate (#45)
PR D of issue #45's umbrella plan. Lifts the V8-coverage-to-Istanbul
range intersection logic into its own publishable crate so any Rust
JS-coverage tool that produces a FileCoverage can fill its hit counts
from V8 inspector ranges without depending on the Oxc instrumenter.

New crate: oxc_coverage_v8 v0.1.0
  - Types: V8FunctionCoverage, V8CoverageRange (serde-compatible with
    node:inspector Profiler.FunctionCoverage so callers can hand the
    V8 inspector output through directly).
  - apply_v8_coverage(file_coverage, source, functions, wrapper_length,
    arm_body_byte_spans): mutates a pre-built FileCoverage by filling
    in statement / function / branch hit counts. Infallible: parse and
    AST-traversal errors are the caller's responsibility.
  - extract_inline_source_map(source) -> Option<serde_json::Value>:
    decodes inline `//# sourceMappingURL=data:application/json[;base64],...`
    trailers (standard + URL-safe base64; percent-encoded fallback).
  - extract_external_source_mapping_url(source) -> Option<&str>: returns
    the trailer URL when it is NOT a data URL (for callers that read
    external `.map` files from disk).
  - Depends on oxc_coverage_types only; zero Oxc / parser deps.

oxc_coverage_instrument changes:
  - v8_to_istanbul.rs slims from 472 to 130 lines: keeps the user-facing
    `v8_to_istanbul` / `v8_to_istanbul_with_loader` orchestrators and the
    `V8ToIstanbulError` enum (the parse failure is specific to this
    crate's AST-traversal pass). The orchestrators call
    `crate::instrument::collect_for_v8_to_istanbul` to build the
    FileCoverage + arm_body_byte_spans side-table, then delegate to
    `oxc_coverage_v8::apply_v8_coverage` for the V8-range intersection,
    then handle inline / external sourceMappingURL extraction via the
    new crate's helpers.
  - Re-exports `V8CoverageRange` and `V8FunctionCoverage` from
    `oxc_coverage_v8` to preserve the existing crate-root public API
    surface byte-for-byte. `V8ToIstanbulError`, `v8_to_istanbul`, and
    `v8_to_istanbul_with_loader` continue to live in the instrument
    crate.

Why split this way (not the full v8_to_istanbul function into the new
crate): the AST-traversal step `collect_for_v8_to_istanbul` is the
instrumenter's responsibility and lives in the instrument crate. Moving
the public `v8_to_istanbul` function into oxc_coverage_v8 would require
that crate to reverse-depend on the instrumenter, which is a build
cycle. The split at `apply_v8_coverage` matches what npm's v8-to-istanbul
actually does internally: parse the source on the caller side, then
apply V8 ranges. The new crate is useful standalone for any Rust tool
that already has its own coverage-map producer.

Release workflow (release-npm.yml):
  - Topological order: types -> source_maps + v8 -> instrument

Verified locally:
  - cargo check / test --workspace --all-targets (passes; 18 test result: ok)
  - cargo test --workspace --doc (passes; instrument 3, source_maps 1,
    types 2, v8 0)
  - cargo clippy --workspace --all-targets -- -D warnings (passes)
  - cargo fmt --all --check (passes)
  - cargo doc --workspace --no-deps --document-private-items -D warnings (clean)
  - cargo deny check (passes)
  - typos (clean)
  - napi build --release + node test.mjs (16/16 pass)
  - scripts/istanbul-upstream-specs.mjs (8/8 pass)
  - scripts/istanbul-diff.mjs (27/27 fixtures match)
  - scripts/real-world-parity.mjs (5/5 libs match)

`cargo publish -p oxc_coverage_v8 --dry-run` fails locally because
oxc_coverage_types is not yet on crates.io; resolves in CI on the next
tag push via the topological publish order.

Refs #45
