+++ crates/oxc-coverage-instrument/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 # `cargo publish` so the lib crate can publish standalone from crates.io.
 oxc_coverage_types = { path = "../oxc-coverage-types", version = "0.1.0" }
 oxc_coverage_source_maps = { path = "../oxc-coverage-source-maps", version = "0.1.0" }
+oxc_coverage_v8 = { path = "../oxc-coverage-v8", version = "0.1.0" }
 oxc_allocator = "0.126"
 oxc_ast = "0.126"
 oxc_codegen = "0.126"
