+++ crates/oxc-coverage-v8/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "oxc_coverage_v8"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+description = "V8 inspector coverage to Istanbul FileCoverage conversion (port of v8-to-istanbul)"
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+categories = ["development-tools::testing", "web-programming"]
+
+[lints]
+workspace = true
+
+[dependencies]
+oxc_coverage_types = { path = "../oxc-coverage-types", version = "0.1.0" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
